+++ kong/clustering/compat/removed_fields.lua
@@ -55,6 +55,7 @@ return {
       "phase_duration_flavor",
     }
   },
+
   -- Any dataplane older than 3.3.0
   [3003000000] = {
     acme = {
@@ -76,6 +77,9 @@ return {
     },
     datadog = {
       "queue",
+    },
+    rate_limiting = {
+      "sync_rate"
     },
   },
 }